EvmKit config: app-owned syncers/decorators, no shared fallback
WalletCore stops self-registering a provider; EvmKitConfigFactory gains a defaultDecorators helper next to defaultSyncers. UnstoppableEvmKitConfigProvider moved into the app and registered in UnstoppableApp.initCore. Each app now configures its own EVM syncers/decorators (stable in StableCore, unstoppable here).
